requirements.txt
# base
django~=2.2
django-markdown-deux
pillow
sorl-thumbnail
python-decouple
django-python3-ldap
ldap3==2.8

# apps.slackbot
requests
djangorestframework
slack_bolt

# apps.discordbot
discord.py
schedule
pyfiglet
cowpy

# Developer tools
pre-commit
